Add ComponentSwitcher
In both the create jam page and the edit jam page we have the same
behaviour of allowing the admin to modify "new" statements. This
consists of the admin triggering an event which causes the statement to
become editable, and from that state triggering some event to go back to
the read-only state.

I added the ComponentSwitcher component so that we can re-use the code
to switch between components without making any assumptions to which
event triggers the switching.